requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
alembic==1.13.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
factory-boy==3.3.0
Faker==20.1.0
